+++ src/transaction/transaction_manager.cpp
@@ -1,13 +1,13 @@
 #include "transaction/transaction_manager.h"
 #include <unordered_set>
 #include <utility>
-#include "common/scoped_timer.h"
+#include "common/timer.h"
 #include "common/thread_context.h"
 #include "metrics/metrics_store.h"
 
 namespace terrier::transaction {
 TransactionContext *TransactionManager::BeginTransaction() {
-  uint64_t elapsed_us = 0;
+  double elapsed_us = 0;
   timestamp_t start_time;
   TransactionContext *result;
   {
@@ -72,7 +72,7 @@ timestamp_t TransactionManager::UpdatingCommitCriticalSection(TransactionContext
 
 timestamp_t TransactionManager::Commit(TransactionContext *const txn, transaction::callback_fn callback,
                                        void *callback_arg) {
-  uint64_t elapsed_us = 0;
+  double elapsed_us = 0;
   timestamp_t result;
   {
     if (common::thread_context.metrics_store_ != nullptr &&
